#197 Unify and improve the NeodymiumConfiguration

+++ src/main/java/com/xceptance/neodymium/util/NeodymiumConfiguration.java
@@ -43,7 +43,7 @@ public interface NeodymiumConfiguration extends Mutable
 
     @Key("neodymium.selenideAddons.staleElement.retry.timeout")
     @DefaultValue("500")
-    public int staleElementRetryTimeout();
+    public long staleElementRetryTimeout();
 
     @Key("neodymium.selenideAddons.optional.retry.pollingIntervall")
     @DefaultValue("3000")
@@ -55,11 +55,11 @@ public interface NeodymiumConfiguration extends Mutable
 
     @Key("neodymium.javaScriptUtils.timeout")
     @DefaultValue("2000")
-    public int javaScriptTimeout();
+    public long javaScriptTimeout();
 
     @Key("neodymium.javaScriptUtils.pollingInterval")
     @DefaultValue("200")
-    public int javaScriptPollingInterval();
+    public long javaScriptPollingInterval();
 
     @Key("neodymium.javaScriptUtils.loading.jQueryIsRequired")
     @DefaultValue("true")
